+++ frontend/src/lib/components/header/AccountSyncIndicator.svelte
